Fixed #144413 and #144588 : partial revert of reading project buildscript.

+++ profiler/src/org/netbeans/modules/profiler/utils/ProjectUtilities.java
@@ -60,7 +60,6 @@ import org.netbeans.modules.profiler.ui.ProfilerDialogs;
 import org.netbeans.spi.java.classpath.PathResourceImplementation;
 import org.netbeans.spi.java.classpath.support.ClassPathSupport;
 import org.netbeans.spi.project.ActionProvider;
-import org.netbeans.spi.project.AuxiliaryConfiguration;
 import org.netbeans.spi.project.SubprojectProvider;
 import org.netbeans.spi.project.support.ant.PropertyEvaluator;
 import org.netbeans.spi.project.support.ant.PropertyUtils;
@@ -79,9 +78,11 @@ import org.openide.util.NbBundle;
 import org.w3c.dom.Element;
 import java.awt.Dialog;
 import java.awt.event.MouseEvent;
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.RandomAccessFile;
 import java.io.UnsupportedEncodingException;
 import java.net.URL;
@@ -102,13 +103,16 @@ import javax.swing.ImageIcon;
 import javax.swing.JButton;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
+import org.netbeans.modules.profiler.projectsupport.utilities.SourceUtils;
 
 
 /**
  * Utilities for interaction with the NetBeans IDE, specifically related to Projects
  *
  * @author Ian Formanek
+ * @deprecated 
  */
+@Deprecated
 public final class ProjectUtilities {
     //~ Static fields/initializers -----------------------------------------------------------------------------------------------
 
@@ -301,7 +305,7 @@ public final class ProjectUtilities {
     }
 
     public static boolean isProfilerIntegrated(Project project) {
-        Element e = project.getLookup().lookup(AuxiliaryConfiguration.class)
+        Element e = ProjectUtils.getAuxiliaryConfiguration(project)
                            .getConfigurationFragment("data", ProjectUtilities.PROFILER_NAME_SPACE, false); // NOI18N
 
         return e != null;
@@ -347,7 +351,11 @@ public final class ProjectUtilities {
     }
 
     public static String getProjectBuildScript(final Project project) {
-        final FileObject buildFile = project.getProjectDirectory().getFileObject("build.xml"); //NOI18N
+        final FileObject buildFile = findBuildFile(project);
+        if (buildFile == null) {
+            return null;
+        }
+
         RandomAccessFile file = null;
         byte[] data = null;
 
@@ -381,6 +389,18 @@ public final class ProjectUtilities {
 
             return null;
         }
+    }
+
+    public static FileObject findBuildFile(final Project project) {
+        FileObject buildFile = project.getProjectDirectory().getFileObject("build.xml"); //NOI18N
+        if (buildFile == null) {
+            Properties props = org.netbeans.modules.profiler.projectsupport.utilities.ProjectUtilities.getProjectProperties(project);
+            String buildFileName = props.getProperty("buildfile"); // NOI18N
+            if (buildFileName != null) {
+                buildFile = project.getProjectDirectory().getFileObject(buildFileName);
+            }
+        }
+        return buildFile;
     }
 
     public static java.util.List<SimpleFilter> getProjectDefaultInstrFilters(Project project) {
@@ -601,10 +621,10 @@ public final class ProjectUtilities {
     }
 
     public static boolean backupBuildScript(final Project project) {
-        final FileObject buildFile = project.getProjectDirectory().getFileObject("build.xml"); //NOI18N
+        final FileObject buildFile = findBuildFile(project);
         final FileObject buildBackupFile = project.getProjectDirectory().getFileObject("build-before-profiler.xml"); //NOI18N
 
-        if (buildBackupFile != null) {
+        if (buildFile != null && buildBackupFile != null) {
             try {
                 buildBackupFile.delete();
             } catch (IOException e) {
@@ -964,10 +984,10 @@ public final class ProjectUtilities {
         FileLock buildBackup2FileLock = null;
 
         try {
-            final FileObject buildFile = project.getProjectDirectory().getFileObject("build.xml"); //NOI18N
+            final FileObject buildFile = findBuildFile(project); //NOI18N
             final FileObject buildBackupFile = project.getProjectDirectory().getFileObject("build-before-profiler.xml"); //NOI18N
 
-            if ((buildBackupFile != null) && buildBackupFile.isValid()) {
+            if (buildFile != null && (buildBackupFile != null && buildBackupFile.isValid())) {
                 try {
                     buildBackupFileLock = buildBackupFile.lock();
 
@@ -1011,7 +1031,7 @@ public final class ProjectUtilities {
         }
 
         // Remove data element from private/private.xml
-        project.getLookup().lookup(AuxiliaryConfiguration.class).removeConfigurationFragment("data", PROFILER_NAME_SPACE, false); // NOI18N
+        ProjectUtils.getAuxiliaryConfiguration(project).removeConfigurationFragment("data", PROFILER_NAME_SPACE, false); // NOI18N
 
         try {
             ProjectManager.getDefault().saveProject(project);
